\name{kuppuswamy_2016}
\alias{kuppuswamy_2016}
\title{Kuppuswamy Socioeconomic Scale Base Data (2016)}
\format{
  A tibble with 7 rows and 3 columns:
  \describe{
    \item{score}{Numeric, score associated with income range.}
    \item{lower}{Numeric, lower income bound.}
    \item{upper}{Numeric, upper income bound.}
  }
}
\description{
  Base income ranges for the Kuppuswamy socioeconomic scale, set for the year 2016 (CPI-IW = 100).
}
\source{
  User-provided data based on Kuppuswamy scale standards.
}
